Fixed bug with Game.java
Sometimes spaceship would not move when user pressed 'w' key. This was due to the fact that inGame = true was set after logic and render thread started. Upon initialization of threads' run function, inGame was false, which caused logic thread to end, and thus spaceship would not update its position.